test(mcp): split HTTP tool tests by tool
